requirements.txt
wandb
tqdm
matplotlib
lightning
torchdyn
ood-metrics
faiss-gpu
torch_geometric
torch_sparse
torch_scatter